fix: 코드 리뷰 반영 — Long 범위 타입 수정, Consumer 파싱 실패 안전 처리
- 1L..totalRequests → 1L..totalRequests.toLong() 타입 불일치 수정 (2곳)
- CouponIssueProcessor: 페이로드 파싱 실패·issueRequest 미발견 시
  EventHandled 저장 후 조기 반환하여 무한 재시도 방지

+++ apps/commerce-api/src/test/kotlin/com/loopers/application/coupon/CouponIssueFacadeIntegrationTest.kt
@@ -211,7 +211,7 @@ class CouponIssueFacadeIntegrationTest @Autowired constructor(
             val successCount = AtomicInteger(0)
 
             // when
-            for (userId in 1L..totalRequests) {
+            for (userId in 1L..totalRequests.toLong()) {
                 executor.submit {
                     try {
                         couponIssueFacade.issueAsync(coupon.id, userId)
